update-approve

+++ backend/Routes/useroom.js
@@ -30,16 +30,64 @@ router.get("/use-room/:id", async (req, res) => {
       `SELECT book_id FROM qrcode WHERE num = :id `,
       { id }
     );
+
     const book_id = qrCheck.rows[0][0];
+    if (!book_id) {
+      return res.status(401).json({ error: "Invalid QR code" });
+    }
+
+    // ดึงข้อมูล startdate จาก booking
+    const bookingCheck = await connection.execute(
+      `SELECT startdate 
+   FROM booking 
+   WHERE book_id = :book_id`,
+      { book_id: book_id }
+    );
+
+    // ตรวจสอบว่า bookingCheck มีข้อมูลหรือไม่
+    if (bookingCheck.rows.length === 0) {
+      return res.status(404).json({ error: "Booking not found" });
+    }
+
+    // ดึง startdate
+    const startdate = bookingCheck.rows[0][0];
+
+    // ตรวจสอบว่ามาใช้ก่อนเวลาเริ่มหรือไม่
+    if (startdate > new Date()) {
+      // ใช้ new Date() เพื่อเปรียบเทียบกับเวลาปัจจุบัน
+      return res
+        .status(403)
+        .json({ error: "Cannot use the room before the start time" });
+    } else if (startdate < new Date()) {
+      return res
+        .status(403)
+        .json({ error: "Unable to use the room because it was past time !!!" });
+    }
+
+    // จากนั้นดำเนินการต่อเพื่อตรวจสอบเวลาที่หมดอายุ
+    const timeCheck = await connection.execute(
+      `SELECT book_id 
+   FROM booking 
+   WHERE book_id = :book_id 
+   AND startdate <= SYSDATE + INTERVAL '30' MINUTE`,
+      { book_id: book_id }
+    );
+
+    const use_id = timeCheck.rows.length > 0 ? timeCheck.rows[0][0] : null;
+
+    if (!use_id) {
+      return res.status(401).json({ error: "Time Out" });
+    }
+
     const deleteQr = await connection.execute(
       `DELETE FROM qrcode WHERE book_id = :book_id `,
-      { book_id: book_id }
+      { book_id: use_id }
     );
     const addUse = await connection.execute(
       `INSERT INTO user_room (book_id, date_use)
          VALUES (:book_id, SYSDATE)`,
       {
-        book_id: book_id,
+        book_id: use_id,
       }
     );
     const status = "SA006";
@@ -48,7 +96,7 @@ router.get("/use-room/:id", async (req, res) => {
       `UPDATE Booking SET app_id = :app_id WHERE book_id = :book_id`,
       {
         app_id: status,
-        book_id: book_id,
+        book_id: use_id,
       }
     );
     await connection.commit();
@@ -59,7 +107,7 @@ router.get("/use-room/:id", async (req, res) => {
         JOIN user_room u ON u.book_id = b.book_id
         WHERE b.book_id = :book_id`,
       {
-        book_id: book_id,
+        book_id: use_id,
       }
     );
     // กำหนดชื่อคอลัมน์ (header) จาก metadata ของคอลัมน์ใน result
